0.3.6
blueprint update
