+++ static/css/_shareElement.sass
@@ -182,6 +182,11 @@
   text-align: center
   margin-top: 10px
   color: #EB5757
+.signupOK
+  width: 100%
+  text-align: center
+  margin-top: 10px
+  color: #27AE60 
 .signinError
   width: 100%
   text-align: center
