+++ static/css/attraction.sass
@@ -38,14 +38,14 @@
         position: absolute
         width: 100%
         bottom: 8px
-        & .dot
-            background-color: $gray300
-            border-radius: 10px
-            +size(8px,8px)
-            margin-left: 4px
-            margin-right: 4px
-            &:hover
-                cursor: pointer
+        // & .dot
+        //     background-color: $gray300
+        //     border-radius: 10px
+        //     +size(8px,8px)
+        //     margin-left: 4px
+        //     margin-right: 4px
+        //     &:hover
+        //         cursor: pointer
 
 .attInfo
     grid-area: B3
