+++ static/css/indexs.css
@@ -336,6 +336,13 @@
   color: #EB5757;
 }
 
+.signupOK {
+  width: 100%;
+  text-align: center;
+  margin-top: 10px;
+  color: #27AE60;
+}
+
 .signinError {
   width: 100%;
   text-align: center;
